Added latest backend changes along with new dashboards-backend feature

+++ packages/backend/dnambc-lib/src/main/java/com/daimler/data/db/jsonb/solution/Solution.java
@@ -27,6 +27,7 @@
 
 package com.daimler.data.db.jsonb.solution;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -104,6 +105,6 @@ public class Solution {
 	private List<SkillSummary> skills;
 	private String dataStrategyDomain;
 	private boolean existingSolution;
-	private Integer requestedFTECount;
+	private BigDecimal requestedFTECount;
 	
 }
